fix: refactor code for sms routes to use common request validation middleware
